pyproject.toml
[project]
name = "cybergym-green-agent"
version = "0.1.0"
description = "CyberGym Green Agent for AgentBeats - Cybersecurity Agent Benchmark"
requires-python = ">=3.11"
dependencies = [
    # A2A and AgentBeats framework
    "a2a-sdk>=0.2.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.30.0",
    "httpx>=0.27.0",
    "loguru>=0.7.0",

    # LLM SDKs
    "openai>=1.0.0",
    "anthropic>=0.40.0",
    "google-genai>=1.0.0",
    "boto3>=1.35.0",
]

[project.scripts]
cybergym-green = "src.green_agent:main"
cybergym-purple = "src.purple_agent:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
